Fix insert sample data accessibility
fix 2
