+++ src/GRA.Web/Areas/MissionControl/Views/Flight/Index.cshtml
@@ -1,6 +1,6 @@
 ﻿<div class="row" style="margin-top: 1rem;">
     <div class="col-sm-3 col-md-2 row-spacing">
-        <a asp-controller="Home" asp-action="LoadSampleData" class="btn btn-info btn-block">Insert sample data</a>
+        <a asp-action="LoadSampleData" class="btn btn-info btn-block">Insert sample data</a>
     </div>
     <div class="col-sm-4 col-md-3 row-spacing">
         <a asp-action="RedeemSecretCodeAsync" class="btn btn-info btn-block">Enter code "secretcode"</a>
